extracted the header into it's own component
