+++ client/src/app/app.component.html
@@ -1,62 +1,3 @@
-<div class="header">
-    <div class="theme" (click)="changeTheme()">
-        <ion-icon
-            [name]="themes[currentTheme].icon + themes.hovered"
-            (mouseout)="themes.hovered = '-outline'"
-            (mouseover)="themes.hovered = ''"
-        />
-    </div>
-    <form class="sort">
-        Sort by
-        <label for="position">
-            <ion-icon 
-                [name]="(todoService.settings.sortingMethod === 'position') ? 'list' : 'list-outline'"
-                title="Sort by position"
-            />
-        </label>
-        <input 
-            type="radio" name="sort" id="position" value="position"
-            [(ngModel)]="todoService.settings.sortingMethod"
-        >
+<app-header/>
 
-        <label for="created">
-            <ion-icon 
-                [name]="(todoService.settings.sortingMethod === 'created') ? 'time' : 'time-outline'"
-                title="Sort by time created"
-            />
-        </label>
-        <input
-            type="radio" name="sort" id="created" value="created"
-            [(ngModel)]="todoService.settings.sortingMethod"
-        >        
-    </form>
-    <div class="settings">
-        <ion-icon name="settings-outline" (click)="toggleSettingsMenu()"/>
-        <div class="settings-box hidden">
-            <form>
-                <label for="confirmRemoval">Require confirmation for todo removals?</label>
-                <input
-                    id="confirmRemoval" name="confirmRemovals" type="checkbox"
-                    [(ngModel)]="todoService.settings.confirmRemovals"
-                >
-                
-                <label for="timeFormat">
-                    How to format time?
-                </label>
-                <select
-                    id="timeFormat" name="timeFormat"
-                    [(ngModel)]="todoService.settings.timeFormat"
-                >
-                    <option value="h24">
-                        24h
-                    </option>
-                    <option value="h12">
-                        12h
-                    </option>
-                </select>
-            </form>
-        </div>
-    </div>
-</div>
-
-<app-todo-list>
+<app-todo-list/>
