+++ client/src/app/app.component.ts
@@ -1,5 +1,4 @@
-import { Component, inject } from '@angular/core';
-import { TodoService } from './todo.service';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -9,41 +8,4 @@ import { TodoService } from './todo.service';
 
 export class AppComponent {
   title = 'getitdone';
-  todoService: TodoService = inject(TodoService);
-
-  currentTheme: 'dark' | 'light' = 'dark';
-  themes = {
-    'hovered': '-outline',
-    'dark': {
-      'icon': 'moon',
-      'colors': {
-        'primary': "#121420",
-        'secondary': "#2C2B3C",
-        'contrast': "#FFFAFA",
-      }
-    },
-    'light': {
-      'icon': 'sunny',
-      'colors': {
-        'primary': "#FFFFFF",
-        'secondary': "#D3D3D3", // lightgrey < vscode color
-        'contrast': "#000000",
-      } 
-    }
-  }
-
-  changeTheme(): void {
-    this.currentTheme = (this.currentTheme === 'dark') ? 'light' : 'dark';
-    const theme = this.themes[this.currentTheme];
-
-    for (let [variable, value] of Object.entries(theme.colors)) {
-      document.documentElement.style.setProperty(`--${variable}`, value);
-    }
-  }
-
-  toggleSettingsMenu() {
-    const settingsMenu = document.getElementsByClassName('settings-box')[0];
-
-    settingsMenu.classList.toggle('hidden');
-  }
 }
