+++ client/src/app/header/header.component.css
@@ -1,0 +1,70 @@
+ion-icon {
+    cursor: pointer;
+}
+
+/* General */
+
+div, form, label {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+div.theme, div.settings {
+    width: 50px;
+    font-size: 36px;
+    padding: 7px;
+}
+
+/* Sorting Bar */
+
+form.sort {
+    flex-grow: 1;
+    gap: 5px;
+    font-size: 24px;
+}
+
+form.sort ion-icon {
+    font-size: 36px;
+}
+
+form.sort > input[type="radio"] {
+    display: none;
+}
+
+/* Settings Box */
+
+div.settings-box {
+    position: absolute;
+    right: 24px;
+    top: 24px;
+
+    padding: 8px;
+    border-radius: 5px;
+    background-color: var(--secondary);
+}
+
+div.settings-box.hidden {
+    display: none;
+}
+
+div.settings-box > form {
+    display: grid;
+    grid-template-columns: auto auto;
+    row-gap: 3px;
+
+    font-size: small;
+}
+
+div.settings-box > form > input {
+    accent-color: var(--contrast);
+}
+
+div.settings-box > form > select {
+    border: 1px solid var(--primary);
+    background-color: var(--secondary);
+}
+
+div.settings-box > form > select > option {
+    color: contrast;
+}
